requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
attrs==23.2.0
backoff==2.2.1
bidict==0.23.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.10.0
dataclasses-json==0.6.6
decorator==4.4.2
defusedxml==0.7.1
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gevent==24.2.1
google-pasta==0.2.0
gradio==4.27.0
gradio_client==0.15.1
greenlet==3.0.3
grpcio==1.64.0
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib_resources==6.4.0
inference-sdk==0.9.17
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proglog==0.1.10
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-magic==0.4.27
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.29
rpds-py==0.18.0
ruff==0.4.1
scikit-image==0.23.2
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
setuptools==69.5.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
supervision==0.20.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tf_keras==2.16.0
thop==0.1.1.post2209072238
tifffile==2024.5.10
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.18
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
